rename Training as Courses
