+++ templates/Courses/index.php
@@ -1,11 +1,11 @@
-<section id="training-page">
+<section id="courses-page">
     <section id="" class="content-section row">
         <?php 
-        $imagePath = str_replace(WWW_ROOT, '', $contentTraining[0]->primary_image_dir) . DS;
+        $imagePath = str_replace(WWW_ROOT, '', $contentCourses[0]->primary_image_dir) . DS;
         $imageURL = str_replace('\\', '/', $imagePath);
         ?>
         <div id="" class="content-image col s12 m5 l6"
-            style="background: url('<?= $imageURL . $contentTraining[0]->primary_image ?>') center center; background-size: cover;">
+            style="background: url('<?= $imageURL . $contentCourses[0]->primary_image ?>') center center; background-size: cover;">
             <div style="background: url('/img/mask_3.png');
                     height: 100%;
                     background-repeat: no-repeat;
@@ -13,8 +13,8 @@
         </div>
         <div id="" class="content col s12 m7 l6">
             <div id="" class="content-text container">
-                <h2><?= $contentTraining[0]->title ?></h2>
-                <p><?= $contentTraining[0]->text ?></p>
+                <h2><?= $contentCourses[0]->title ?></h2>
+                <p><?= $contentCourses[0]->text ?></p>
             </div>
         </div>
     </section>
@@ -49,4 +49,4 @@
 
     <?= $this->element('reinforcement') ?>
 
-</section>  
+</section>
