+++ webroot/scss/main.scss
@@ -498,7 +498,7 @@ nav {
     }
 }
 
-#training-page {
+#courses-page {
     #step-by-step-section {
         width: 75%;
 
@@ -636,7 +636,7 @@ footer {
         }
     }
 
-    #training-page {
+    #courses-page {
         #step-by-step-section {
             width: 85%;
         }
